consolidate helpers with index
